Update
Included a Strategy design pattern for password strength checking depending on a user's role (Admin requires strongest password). Got rid of unnecessary unit tests. Created a User Factory class for creating a user (depending on their role )in the web space because it is an abstract object and the web environment does not support it.